Backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.4.2
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
openai==1.6.1
pandas==2.0.3
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
zipp==3.17.0
